+++ tests/Utils/Psr7/DummyResponse.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace WoohooLabs\Harmony\Tests\Utils\Psr7;
 
 use InvalidArgumentException;
+use Laminas\Diactoros\Stream;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\StreamInterface;
-use Zend\Diactoros\Stream;
 
 use function is_array;